Reference

PeriodicArrays.PeriodicArrays — Module

Arrays with fixed size and arbitrary boundary conditions.

PeriodicArrays.PeriodicArray — Type

PeriodicArray{T, N, A, F} <: AbstractArray{T, N}

N-dimensional array backed by an AbstractArray{T, N} of type A with fixed size and periodic indexing as defined by map.

array[index...] == map(array[mod1.(index, size)...], fld.(index .- 1, size)...)

PeriodicArrays.PeriodicArray — Method

PeriodicArray(data, [map])

Create a PeriodicArray backed by data. map is optional and defaults to the identity map.

PeriodicArrays.PeriodicArray — Method

PeriodicArray(def, size, [map])

Create a PeriodicArray of size size filled with value def. map is optional and defaults to the identity map.

PeriodicArrays.PeriodicMatrix — Type

PeriodicMatrix{T, A, F} <: AbstractMatrix{T}

Two-dimensional array backed by an AbstractArray{T, 2} of type A with fixed size and periodic indexing. Alias for PeriodicArray{T, 2, A, F}.

PeriodicArrays.PeriodicVector — Type

PeriodicVector{T, A, F} <: AbstractVector{T}

One-dimensional array backed by an AbstractArray{T, 1} of type A with fixed size and periodic indexing. Alias for PeriodicArray{T, 1, A, F}.

array[index] == map(array[mod1(index, length)], fld(index - 1, length))